Made VuforiaLocaliztionTest
This program acts like the usual LocalizationTest, but instead of keeping track of movement with odometers, it uses the vuforia localization to produce its actual location. It will only localize when holding B, and the robot will automatically stop itself when doing do. After this, pressing A and Y simultaneously will have the robot generate a trajectory and follow it autonomously. It will use the drive's pose estimate (which is influenced by vuforia after holding B) and attempt to drive to the specified Waypoint, currently defined as the center of the field.